requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.30.6

# Database & ORM
sqlalchemy==2.0.36      


# Authentication & Security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
argon2-cffi==23.1.0
email_validator==1.3.1
# Data Validation
pydantic==2.9.2

# Environment Variables
python-dotenv==1.0.1

# Scheduling (for automated payroll runs)
apscheduler==3.10.4

# Reporting & Exports
reportlab==4.2.2       # PDF reports
openpyxl==3.1.5        # Excel exports
pandas==2.2.3

# Utilities
requests==2.32.3

# Testing
pytest==8.3.2
httpx==0.27.2          # async test client
